refactor: Alter user-repository folder to user folder in infrastructure
